Refactor navigation structure by replacing Nav component with TopNav and BottomNav for improved layout. Remove deprecated loading component and integrate loading state directly into SearchPage. Update database schema to track game submissions with a new foundBy field.
