--- src/app/search/loading.tsx
@@ -1,42 +1,0 @@
-import { Card, CardContent } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-import Link from "next/link";
-import { ArrowLeft } from "lucide-react";
-
-export default function SearchLoading() {
-  return (
-    <div className="container max-w-5xl mx-auto py-6 px-4">
-      <div className="mb-8">
-        <Link
-          href="/"
-          className="inline-flex items-center text-sm text-muted-foreground hover:text-foreground mb-4"
-        >
-          <ArrowLeft className="mr-1 h-4 w-4" />
-          Back to Discover
-        </Link>
-
-        <h1 className="text-3xl font-bold mb-6">Search Games</h1>
-
-        <div className="flex gap-2">
-          <Skeleton className="h-10 flex-1" />
-          <Skeleton className="h-10 w-24" />
-        </div>
-      </div>
-
-      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
-        {Array(6)
-          .fill(0)
-          .map((_, i) => (
-            <Card key={i} className="h-full overflow-hidden">
-              <Skeleton className="h-[215px] w-full" />
-              <CardContent className="p-4">
-                <Skeleton className="h-6 w-4/5 mb-2" />
-                <Skeleton className="h-4 w-full" />
-                <Skeleton className="h-4 w-2/3 mt-1" />
-              </CardContent>
-            </Card>
-          ))}
-      </div>
-    </div>
-  );
-}
